requirements.txt
typer
pdfplumber
chromadb
sentence-transformers
openai
tqdm
numpy
python-dotenv
fastapi
uvicorn
tf-keras
hf_xet
chroma-migrate